requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
python-docx==1.1.0
reportlab==4.0.9
google-generativeai==0.3.2
aiofiles==23.2.1
